+++ book_tools/format/epub.py
@@ -5,7 +5,7 @@ from tempfile import mktemp
 from book_tools.format.aes import encrypt
 from book_tools.format.bookfile import BookFile
 from book_tools.format.mimetype import Mimetype
-from book_tools.format.util import list_zip_file_infos
+from book_tools.format.util import list_zip_file_infos, safe_lxml_parser
 
 class EPub(BookFile):
     class Issue(object):
@@ -88,7 +88,7 @@ class EPub(BookFile):
     def __etree_from_entry(self, info):
         with self.__zip_file.open(info) as entry:
             try:
-                return etree.fromstring(entry.read(1048576))
+                return etree.fromstring(entry.read(1048576), parser=safe_lxml_parser())
             except:
                 raise EPub.StructureException('\'' + info.filename + '\' is not a valid XML')